Fixed diagonal speed to normal speed, fixed boundaries of the window

+++ src/entity/Entity.java
@@ -5,7 +5,7 @@ import java.awt.image.BufferedImage;
 public class Entity {
 	public int x, y;
 	
-	public int speed;
+	public double speed;
